Ajustes ponteiro pra função
