+++ src/AntiColisao.cpp
@@ -1,5 +1,7 @@
+#include <stack>
 #include "AntiColisao.hpp"
 #include "Helper.hpp"
+#include "RoboCore_Vespa.h"
 
 VespaServo servo;
 VespaMotors motores;
@@ -14,26 +16,32 @@ void setup_anti_colisao() {
 }
 
 void loop_anti_colisao() {
+    // Ponteiro pra função
+    int (*pSensor)() = &sensor_ultrassonico;
+    void (*pVerificaObstaculos)(VespaServo &, std::stack<uint_fast8_t> &) = &verificaObstaculos;
+    
+    // Pilha 
     std::stack<uint_fast8_t> pilha{};
-    int distancia;
+
+    int distancia  = 0;
     
     while(true) 
     {
         delay(ESPERA); //aguarda o tempo de espera para leitura do sensor
-        distancia = sensor_ultrassonico(); //armazena a distancia lida pelo sensor a variavel distancia
+        distancia = pSensor(); //armazena a distancia lida pelo sensor a variavel distancia
 
         //verifica se a distancia lida pelo sensor e menor ou igual ao valor configurado na variavel "DISTANCIA_OBSTACULO"
         if (distancia <= DISTANCIA_OBSTACULO) //se for verdadeiro
         {            
             delay(ESPERA); //aguarda o tempo de espera para leitura do sensor
-            distancia = sensor_ultrassonico(); //atualiza a leitura do sensor
+            distancia = pSensor(); //atualiza a leitura do sensor
 
             //confirma se a distancia lida pelo sensor e menor ou igual ao valor configurado na variavel "DISTANCIA_OBSTACULO"
             if (distancia <= DISTANCIA_OBSTACULO) 
             {
                 motores.stop();
                 delay(ESPERA);          
-                verificaObstaculos(servo, pilha);
+                pVerificaObstaculos(servo, pilha);
 
                 // Enquanto tiver 2 obstáculos dos lados, continua andanda para trás
                 while (pilha.size() == 2) {
@@ -41,7 +49,7 @@ void loop_anti_colisao() {
                     pilha.pop();
                     andaParaTras(motores, ESPERA_MOVIMENTO);
                     delay(ESPERA);
-                    verificaObstaculos(servo, pilha);
+                    pVerificaObstaculos(servo, pilha);
                 }
 
                 // Se a pilha estiver vazia, ou seja, sem obstáculos dos lados
@@ -77,49 +85,4 @@ void loop_anti_colisao() {
             motores.forward(VELOCIDADE); //mantem o robo andando para frente
         }
     }
-}
-
-// Gira o robô para os lados e verifica se tem obstáculos 
-void verificaObstaculos(VespaServo &servo, std::stack<uint_fast8_t> &pilha) 
-{
-    int distancia;
-
-    // Gira Sensor para direita
-    servo.write(0);
-    delay(ESPERA_GIRO_SENSOR);
-    distancia = sensor_ultrassonico();
-    
-    if (distancia <= DISTANCIA_OBSTACULO)
-    {
-        pilha.push(OBSTACULO_DIREITA);
-    }
-
-    delay(ESPERA);
-    
-    // Gira Sensor para esquerda
-    servo.write(180);
-    delay(ESPERA_GIRO_SENSOR);
-    distancia = sensor_ultrassonico();
-    
-    if (distancia <= DISTANCIA_OBSTACULO)
-    {
-        pilha.push(OBSTACULO_ESQUERDA);
-    }
-
-    // Volta pra o centro
-    delay(ESPERA_GIRO_SENSOR);
-    servo.write(90);    
-}
-
-//funcao para a leitura do sensor
-int sensor_ultrassonico() 
-{
-    //realiza o pulso de 10 microsegundos no trigger do sensor
-    digitalWrite(PINO_TRIGGER, HIGH);
-    delayMicroseconds(10);
-    digitalWrite(PINO_TRIGGER, LOW);
-
-    //mede o pulso em microsegundos retornado para o echo do sensor
-    //e converte o tempo para distancia divindo por 58
-    return pulseIn(PINO_ECHO, HIGH) / 58;
 }
